// web-application/src/types-openapi/api/models/HeatmapCellDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HeatmapCellDto
 */
export interface HeatmapCellDto {
    /**
     * 
     * @type {number}
     * @memberof HeatmapCellDto
     */
    dayOfWeek?: number;
    /**
     * 
     * @type {number}
     * @memberof HeatmapCellDto
     */
    hour?: number;
    /**
     * 
     * @type {number}
     * @memberof HeatmapCellDto
     */
    bookings?: number;
}

/**
 * Check if a given object implements the HeatmapCellDto interface.
 */
export function instanceOfHeatmapCellDto(value: object): value is HeatmapCellDto {
    return true;
}

export function HeatmapCellDtoFromJSON(json: any): HeatmapCellDto {
    return HeatmapCellDtoFromJSONTyped(json, false);
}

export function HeatmapCellDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): HeatmapCellDto {
    if (json == null) {
        return json;
    }
    return {
        
        'dayOfWeek': json['dayOfWeek'] == null ? undefined : json['dayOfWeek'],
        'hour': json['hour'] == null ? undefined : json['hour'],
        'bookings': json['bookings'] == null ? undefined : json['bookings'],
    };
}

export function HeatmapCellDtoToJSON(json: any): HeatmapCellDto {
    return HeatmapCellDtoToJSONTyped(json, false);
}

export function HeatmapCellDtoToJSONTyped(value?: HeatmapCellDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dayOfWeek': value['dayOfWeek'],
        'hour': value['hour'],
        'bookings': value['bookings'],
    };
}