// web-application/src/types-openapi/api/models/RecentTransactionDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RecentTransactionDto
 */
export interface RecentTransactionDto {
    /**
     * 
     * @type {string}
     * @memberof RecentTransactionDto
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RecentTransactionDto
     */
    type?: string | null;
    /**
     * 
     * @type {number}
     * @memberof RecentTransactionDto
     */
    amount?: number;
    /**
     * 
     * @type {Date}
     * @memberof RecentTransactionDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof RecentTransactionDto
     */
    customerName?: string | null;
}

/**
 * Check if a given object implements the RecentTransactionDto interface.
 */
export function instanceOfRecentTransactionDto(value: object): value is RecentTransactionDto {
    return true;
}

export function RecentTransactionDtoFromJSON(json: any): RecentTransactionDto {
    return RecentTransactionDtoFromJSONTyped(json, false);
}

export function RecentTransactionDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecentTransactionDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'type': json['type'] == null ? undefined : json['type'],
        'amount': json['amount'] == null ? undefined : json['amount'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'customerName': json['customerName'] == null ? undefined : json['customerName'],
    };
}

export function RecentTransactionDtoToJSON(json: any): RecentTransactionDto {
    return RecentTransactionDtoToJSONTyped(json, false);
}

export function RecentTransactionDtoToJSONTyped(value?: RecentTransactionDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'type': value['type'],
        'amount': value['amount'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'customerName': value['customerName'],
    };
}