// web-application/src/types-openapi/api/models/TopCourtDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TopCourtDto
 */
export interface TopCourtDto {
    /**
     * 
     * @type {string}
     * @memberof TopCourtDto
     */
    courtId?: string;
    /**
     * 
     * @type {string}
     * @memberof TopCourtDto
     */
    courtName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof TopCourtDto
     */
    bookingCount?: number;
}

/**
 * Check if a given object implements the TopCourtDto interface.
 */
export function instanceOfTopCourtDto(value: object): value is TopCourtDto {
    return true;
}

export function TopCourtDtoFromJSON(json: any): TopCourtDto {
    return TopCourtDtoFromJSONTyped(json, false);
}

export function TopCourtDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TopCourtDto {
    if (json == null) {
        return json;
    }
    return {
        
        'courtId': json['courtId'] == null ? undefined : json['courtId'],
        'courtName': json['courtName'] == null ? undefined : json['courtName'],
        'bookingCount': json['bookingCount'] == null ? undefined : json['bookingCount'],
    };
}

export function TopCourtDtoToJSON(json: any): TopCourtDto {
    return TopCourtDtoToJSONTyped(json, false);
}

export function TopCourtDtoToJSONTyped(value?: TopCourtDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'courtId': value['courtId'],
        'courtName': value['courtName'],
        'bookingCount': value['bookingCount'],
    };
}